# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MiruroAPI",
    "description": "Free RESTful API for anime streaming data. Powered by AniList GraphQL and Miruro streaming providers.",
    "version": "2.2.0",
    "contact": {
      "name": "Shineii86",
      "url": "https://github.com/Shineii86/MiruroAPI"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "https://mirurotvapi.vercel.app",
      "description": "Production"
    }
  ],
...
  "paths": {
    "/api/search": {
      "get": {
        "tags": ["Search"],
        "summary": "Search anime by name",
        "parameters": [
          { "name": "query", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "per_page", "in": "query", "schema": { "type": "integer", "default": 20, "maximum": 50 } }
        ]
      }
    },
    "/api/suggestions": {
      "get": {
        "tags": ["Search"],
        "summary": "Autocomplete suggestions",
        "parameters": [
          { "name": "query", "in": "query", "required": true, "schema": { "type": "string" } }
        ]
      }
    },
...
    "/api/sources": {
      "get": {
        "tags": ["Streaming"],
        "summary": "Streaming sources (detailed)",
        "parameters": [
          { "name": "episodeId", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "provider", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "anilistId", "in": "query", "required": true, "schema": { "type": "integer" } },
          { "name": "category", "in": "query", "schema": { "type": "string", "default": "sub" } }
        ]
      }
    },
...
  }
}